Refactored homescreen->separate market page
