+++ pages/dashboard.js
@@ -1,42 +1,12 @@
-import React, { useState } from 'react';
-import { Text, View, StyleSheet, TextInput, FlatList } from 'react-native';
+import React from 'react';
+import { StyleSheet } from 'react-native';
 
-import { createNativeStackNavigator } from '@react-navigation/native-stack';
 import { createBottomTabNavigator } from '@react-navigation/bottom-tabs';
 
-import ItemDetail from './ItemDetail'; // Import the ItemDetail component
-import ProfilePage from './ProfilePage'; // Import the ProfilePage component
+import Market from './Market'; // Import the ItemDetail component
+import ItemPage from './ItemPage'; // Import the ProfilePage component
 
-const Stack = createNativeStackNavigator();
 const Tab = createBottomTabNavigator();
-
-function Home() {
-    // Contains the stack navigator for the home screen to allow for navigation to the ItemDetail page
-    return (
-        <Stack.Navigator initialRouteName="Home" screenOptions={{ headerShown: false }}>
-            <Stack.Screen name="Home" component={HomeScreen} />
-            <Stack.Screen name="ItemDetail" component={ItemDetail} />
-        </Stack.Navigator>
-    );
-}
-
-
-
-const HomeScreen = () => {
-    const [searchText, setSearchText] = useState('');
-
-    return (
-        <View style={styles.container}>
-            <TextInput
-                style={styles.searchInput}
-                placeholder="Search for items..."
-                value={searchText}
-                onChangeText={setSearchText}
-            />
-            <Text style={styles.dummyText}>Items will be added upon backend creation</Text>
-        </View>
-    );
-};
 
 export default function Dashboard({ route }) {
     // Extract the username from the navigation route parameters
@@ -44,10 +14,10 @@ export default function Dashboard({ route }) {
 
     return (
         <Tab.Navigator>
-            <Tab.Screen name="Market" component={HomeScreen} />
+            <Tab.Screen name="Market" component={Market} />
             {/* Pass the username to the ProfilePage */}
-            <Tab.Screen name="Profile">
-                {() => <ProfilePage username={username} />}
+            <Tab.Screen name="Items">
+                {() => <ItemPage username={username} />}
             </Tab.Screen>
         </Tab.Navigator>
     );
